separate logic of components
