+++ src/environments/environment.ts
@@ -1,4 +1,5 @@
 export const environment = {
 	apiUrl: 'https://eyeland-api-production.up.railway.app/api/teachers',
+	// apiUrl: 'http://localhost:3000/api/teachers',
 	socketUrl: 'https://eyeland-api-production.up.railway.app/teachers'
 };
