+++ src/shared/enums/Socket.enum.ts
@@ -1,4 +1,5 @@
 export enum SocketEvents {
+	JOIN = 'join',
 	TEAMS_STUDENT_UPDATE = 'teams:student:update',
 	TEAMS_STUDENT_ANSWER = 'team:student:answer',
 	COURSE_LEADERBOARD_UPDATE = 'course:leaderboard:update'
